requirements.txt
numpy
pandas
matplotlib
pillow
python-dotenv
seaborn